Refactor Equip methods

+++ UdpHosts/GameServer/Enums/ItemDynamicFlags.cs
@@ -5,7 +5,7 @@ namespace GameServer.Enums;
 
 [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1602:EnumerationItemsMustBeDocumented", Justification = "TODO")]
 [Flags]
-public enum ItemDynamicFlags : uint
+public enum ItemDynamicFlags : byte
 {
     IsBound = 0x01,
     Unk_0x02 = 0x02, // is_new?
